+++ docs/CHANGELOG.md
@@ -5,6 +5,10 @@
       - Added option **--cphash** to output the cpHash for the command
         TPM2_CC_ECC_Parameters.
 
+  * tpm2_flushcontext:
+      - Added option **--cphash** to output the cpHash for the command
+        TPM2_CC_FlushContext.
+
   * tpm2_ecephemeral:
       - Added option **--cphash** to output the cpHash for the command
         TPM2_CC_EC_Ephemeral.